Code adjustments and futureproofed debug
Probably not much, and maybe not pretty on the '2D on 3D' cheat lens. But aim wise, it at least ensures the turret rotates to nearest enemy, without constantly calling for nearest enemy all the time.

Might be unrealistic, but less likely to be a 'RAM bottleneck', under avoided "foreach" call per EVERY update lens.